==> 01. Time Series Data.ipynb <==
# ipynb是python Notebook, now is called Jupyter Notebook more

# Introduction to Time Series Data

Data can come in many different formats, and many differentshapes and sizes. You've maybe heard of tabular data, a format you may be familiar with from working in something like Excel. 

We will explore two main kinds of tabular data in this module. The first is time series data. Time series data will be *indexed* with a date and time. We'll look a bit more closely at that soon, but for now just think of it as each row having a date or time, rather than a row number.

## Loading Data

One of the most popular packages in Python for working with tabular data is called Pandas. Today we'll get acquainted with Pandas.

The first thing we'll do is `import` the `pandas` package. Convention has us use a shortform name - `pd` - because we'll be using the package so often.

# Python 本身是个工具箱，但处理表格数据这个“螺丝”，它没有现成的“螺丝刀”
# pandas 就是这个螺丝刀，专门用来搞定数据分析相关的工作，用于load data

import pandas as pd

# pandas's shortform is pd

And below we'll use pandas' `read_csv()` to load the data into a `DataFrame`. DataFrames are the main data structure in pandas for tabular data, and lots of other programming languages use the concept of a DataFrame too! By convention, you'll often see `df` used as a variable name.

df = pd.read_csv("data/NVDA_2024.csv")
# df is dataframe, the above means I dowloaded the tips dataset from pandas(pd) with its tool read_csv
# with data set NVDA_2024.csv from data file

Before we do anything else, it's a good idea to take a look at the DataFrame. Some methods will let us take a closer look at parts of our data. 

df
# 查看全部data frame
df.head()
df.tail()
# 查看头和末尾几行（默认是5行）
print(df.head())
# 把内容“显示在屏幕上”

## Looking at data

Other methods and attributes can give us an overview, and give us further insights to our data in general. `shape()` will tell us the number of rows and columns in our data frame, while `info()` will give us some info on the data type (`dtype`) of each column.

You'll notice the types are slightly different from the usual Python types - this is because they belong to the `numpy` package, which sits under the hood of `pandas`. We'll look more at `numpy` tomorrow, but for now here is a word about each of the types in our data frame.

- `float64` - 64-bit floating point (number with a decimal point)
- `int64` - 64-bit integer (whole number)
- `object` - other Python data types (strings in this case)

df.shape
# 告诉信息多少row多少column
df.info()
# 说数据的数量、有无空信息，数据的type
# object is text, float is decimal number, int is integer
# 各种code之后的符号怎么用？
    # 什么都不加是属性（信息），变量本身
    # () 是调用方法，要做的事情
    # [] 是索引，取值


For a look at some actual data within the data frame, we can use square bracket notation and `iloc` to access columns and rows. The `i` in `iloc` refers to **integer-based indexing**, so looking at a row or column *number*.

df["Close"]
# 取出Close这一列的内容
df[["Open","Close"]]
# 取出Open和Close这两列的内容，因为有两个数据所以要[[]]

df.iloc[0]
# iloc是integer/index location, 用“整数编号位置”取行/列，[0] 表示取第 0 行
df.iloc[0,2]
# 0行，2列的单元格的值
# 还有一个df.loc，是按数据本身的“标签名字”取行或列

## Filtering Data

In Pandas, we can use a technique known as *boolean indexing* or *masking* to filter rows depending on some condition. We can express conditions using a *boolean expression* or *compound boolean expression* with either `&` (and) *or* `|` (or). These are also called *logical expressions*.

condition = df["Date"] == "2024-08-08"
# 条件判断，用来筛选出“Date列等于 2024-08-08 的行”
# == for equal, = for formular assignment
df[condition]


df[(df["Date"] >= "2024-08-01") & (df["Date"] <= "2024-08-31")]
# 写很多次df["Date"]的原因是：每个条件（左边和右边）都在独立判断 Date 列，必须写全
# 最外层的df[]是 Pandas 的“筛选行”操作

### Exercise: End of Year

Display the data for the entire month of December 2024.

df[df["Date"] >= "2024-12-01"] 

### Exercise: The First Fifty

Display the data for the first fifty (50) days of trading in the period.

df.iloc[0:50]
# 注意我之前跑不动是因为用了()而不是[]，()是函数调用，[]是位置索引

### Optional Exercise: Big Days

Display rows where trading volume exceeded 800 000 000.

df[df["Volume"] > 800_000_000]

## Setting the Index

In a DataFrame, each row is assigned a unique index value. By default, this is just a number (starting at 0). When it makes sense, we can choose one of the other columns to be an index. For time series data, where each row represents a different point in time, we'll set our `Date` column as the index. This will make it easier for us to work with the data, and can speed up other operations later on.


df["Date"] = pd.to_datetime(df["Date"])
# pd.to_datetime()是Pandas的一个函数，把传进去的内容转换成标准的“datetime object”格式
# df["Date"]=是把转换后的日期时间数据重新存回df的"Date"这一列，以覆盖原来的内容
df.set_index("Date")
# df.set_index("Date")：把"Date"这一列，设置成新的行索引（index），去掉原来的数字行号0,1,2,3……
# 用“Date”列做行索引，就是用你数据本身有意义的日期标签来定位每行；如可以用df.loc[2025-01-02],当然.iloc依然还可以用
# 如果不设置索引，访问数据时只能用数字下标（比如iloc），要用.loc也是只能写df.loc[0]（没额外用）；set index后可以用日期筛选

df
# 但是，python非常保守，除非specifically要求他这样干，他不会改数据
# 如，运行set_index之后只会print preview，而不会真的去改动df；运行一下df看看吧，没有改动

df["Date"] = pd.to_datetime(df["Date"])
df.set_index("Date", inplace=True)
# inplace=True：表示直接在原来的 df 上修改，不返回新表；就是让之前的set index的命令改df
df
# 注意，这个cell的命令跑完之后，再跑二次会发现不行，会出现error
# 仔细看error出现的原因（always从下往上看，最后的原因往往是最重要的）是Date，意思是python找不到叫Date的column了
# 为什么找不到了？因为把Date那个column升格为“行标签”index了
# 如何重新弄出来？点菜单的run all

We convert the 'Date' column to a datetime object because pandas can recognise and efficiently work with datetime objects. We set the `Date` column as the index because in time-series data like ours, operations are time-based.

With the index set, we can now use it to access different portions of our data a little bit more easily. Because our indices are labeled, we can use `loc` for **label-based indexing**.

# 好，现在就可以用日期做index来locate数据了
df.loc["2024-08"]
# print(df.loc['2024-08'])

df.loc["2024-01-01" : "2024-01-07"]
# print(df.loc["2024-08-08":"2024-08-12"])

df.loc["2024-08-08", "Close"]
# But it is preferable to use .at for a single cell
# print(df.at['2024-08-08', 'Close'])



## Getting the Index

Oftentimes it is helpful to retrieve the index of the dataframe for a given row or rows. Let's say we wanted to see the dates where Nvidia's `Volume` was less than 150 000 000. After some smart boolean indexing, we can look at the `index` attribute.

dates = df[df["Volume"] < 150_000_000].index
# 找出成交量小于 1.5 亿的所有行，并提取这些行的索引（日期）
# .index取出这些行对应的索引（set index之后就是日期了，否则是数字行号）
# dates = 是把这些索引结果赋值给变量 dates
dates

Because of our conversion to datetime when we first loaded the dataframe, we can pull specific information out of the index.

dates.day_name()
# 如果dates（上个命令生成的）是一个 DatetimeIndex, 那么.day_name()给这些日期对应的星期几
# 注意.day_name是function，所以需要加()

dates.to_period("Q")
# 给这些日期对应是哪个quater

df.info()
# 用于确认现在的index是什么index，如现在是DatetimeIndex


## Aggregations

There are many basic operations we can do with pandas, such as calculating the mean of a column, the maximum of a column, and so on. We generally refer to these as *aggregations* since they reduce multiple values to one summary value.


df.Close
# 是df["Close"]的简写，不用[]和""了
# 不要用简写的情况：1.column name有空格；2.第一次创建一个column；3.column name是python自己的功能（如head，tail）

df.Close.mean()
# the mean of close column of the whole preiod
df.Volume.max()
# volume最大值
df.Volume.idxmax()
# volume最大值出现的date（give me the index）

df[["Low", "High"]].mean()
# the mean of each of multiple columns
df[["Low", "High"]].mean(axis=1)
# the mean of low and high together of each row

### Exercise: nVidia Quarters

Did Q3 or Q4 have more trading days where the `Close` price was above the annual average (i.e. above the mean)?

condition = df.Close > df.Close.mean()
mask = condition.index.to_period("Q") == "2024Q3"
# mask = 是保存结果到变量；== "2024Q3"：判断每个日期是不是在 2024Q3

print(condition.loc[condition.index.to_period("Q") == "2024Q3"].sum())
print(condition.loc[condition.index.to_period("Q") == "2024Q4"].sum())

### Exercise: March Madness

Looking only at the month of March, print the following information:

* First opening price of the period
* Last close price of the period
* Total volume traded over the period

# 我的方法：
print(df.loc["2024-03-01","Open"])
print(df.loc["2024-03-28","Close"])

# 老师的方法：
march_data = df.loc["2024-03"]
march_data

march_data.Open.iloc[0]
march_data.Close.iloc[-1]
march_data.Volume.sum()

# 想要保存文档，打开terminal（下面那个窗口），输入：
# jupyter nbconvert --to html (文档的名字，tab自动跳完) 【然后点enter】
# 给一个html，下载，打开就是个网页可以download了
