==> 02. Data Cleaning.ipynb <==
## Data Cleaning

Let's load the some data for TSLA. Unfortunately, this data is not quite as *clean* as our NVDA data, so we'll need to do some data wrangling. The file we're looking to load is `TSLA_2015_2024.csv`.

import pandas as pd

df = pd.read_csv("data/TSLA_2015_2024.csv")
df

df.Date = pd.to_datetime(df.Date, format="%d/%m/%Y")
# 把传进去的内容转换成标准的“datetime object”格式，出现了个error，format有问题，于是改format
df.set_index("Date", inplace=True)
# 设置为新的index
df

Can you see what we mean by messy? How many issues can you spot?

- Dates out of order
- Duplicate rows
- Missing values

## Ordering and Duplicates

First let's start with sorting the index.

df.index.is_monotonic_increasing
# 用于检查数据是不是monotonic的increasing，给答案不是

df.sort_index(inplace = True)
# sort一下，然后inplace保证确实改动
df

df.index.is_monotonic_increasing
# 现在给答案，是

Now let's focus on duplicates:

df.duplicated()
# 给每个row展示有没有duplicate
df.duplicated().sum()
# false=0, true=1,这个展示有多少被duplicate

df.drop_duplicates(inplace=True)
df

df.duplicated().sum()
# 现在duplicate的是0个了

#### Tip: Method Chaining

**Method chaining** is a popular feature of pandas. It allows us to *chain* together several operations in a single line of code. For example, we can set the index, sort the data frame and drop any duplicates all at once. Notice we don't use `inplace` but rather re-assign to the original `df` variable.

```python
df = df.set_index("Date").sort_index().drop_duplicates()
```

# 一次性完成时间、sort和去除duplicates的code：
# df = df.set_index("Date", format="%d/%m/%Y").sort_index().drop_duplicates()
# 其中format=是看有没有这个format的问题而导致跑不动再加，没有的话可以不加

## Not a Number (NaN)

### Exercise: Some Null Chain

Let's look at the missing or `NaN` values next. Previously, we saw that `info()` gave us some insight into how many missing values we had, but we can also use `isnull()`.

Can you chain `isnull()` with `sum()` to get a single value stating the total number of missing values in the data frame?

df.isnull().sum()
# 给每个column的null的sum

df.isnull().sum().sum()
# 给全部column的null的sum

df.isnull()

We can find out which rows have missing data using `isnull()`, `any()` along rows and some smart *masking*.

condition = df.isnull().any(axis=1)
df[condition]

Now that we've identified our missing values, the big question is how to handle them. There are many approaches to this that will vary depending on the data and the further analysis you plan to carry out.

# df.dropna()
# na is for not available/null; drop all na

df.dropna(subset="Close")
# drop na only in Close column

df.Volume.fillna(df.Volume.mean())
# fill na with Volume.mean

df.Close.interpolate(method = "linear")
# fill na with“前后的数值”所算出中间的合理值;"linear"：线性插值

df.Close.ffill()
# fill na with the previous number

# 以上的命令都没有执行，因为没有inplace；如果运行错了命令，直接删除错的命令，然后run all就好


### Exercise: Cleaning up

Notice how above we didn't actually update the `df` variable, so our DataFrame is still full of missing values. Fix all missing values applying the following rules:
- Fill missing Close by linear interpolation
- Fill missing Volume with the value from the day before
- Fill missing Open with the median Open
- Fill missing High with the Close or Open, whichever is higher
- Fill missing Low with a value 3% lower than the High


Your DataFrame `df` should have no missing values when done. Use `info()` to confirm.

**NOTE:** When changing values in a data frame, it is recommended to avoid using `inplace`, and instead re-assign the variable.

df.Close = df.Close.interpolate(method = "linear")
# 用后面那个来overwrite
df.Volume = df.Volume.ffill()
df.Open = df.Open.fillna(df.Open.median())

df.High = df.High.fillna(df[["Close","Open"]].max(axis=1))
# (axis=1)单独看每个row哪个更高
df.Low = df.Low.fillna(df.High * 0.97)
# 直接*

df.isnull().sum().sum()

#### Advanced: Data Types

You may have noticed that the **Volume** column in the 2021 data frame is a `float64` instead of the `int64` *dtype* we had in the 2020 data frame. Missing values (NaN) are represented as a special case of floating point number, so all the values in **Volume** were automatically *upcast* to floats.

Ideally our columns should be of the *dtype* that most accurately represents them. This will improve performance when working with large data frames. Now that we've resolved our missing numbers, we can *cast* our trading volumes as integers.

df.info()
df.Volume = df.Volume.astype("int64")
# 把volume列的数据类型转换为整数型int64，int64是64位整数，能表示非常大的整数值（适合金融数据中大成交量）
df.info()

## Saving Data

Now that we've cleaned our data, let's save it, by writing it to a new .CSV file. We can use pandas' `to_csv()`.

df.to_csv("data/TSLA_clean.csv")
# 现在这个就在data的下面保存了这个文档了

df
# 可以点生成出来的表格的右下角带三个arrow的表格，给出一个新tab展示数据信息，每个阶段都可以点

# inplace和=都是起到对数据库进行修改的动作，inplace比较old school
# 老师的做法是：单命令用inplace，多命令在一条用=，如：
# df.sort_index(inplace=True) 
# df = df.set_index("Date").sort_index().drop_duplicates()
