==> 03. Data Analysis.ipynb <==
# Data Analysis

Now that we've got clean data, let's start with some basic financial analysis.

First, let's load our CSV file into a DataFrame, covert our dates, set the index, and check for duplicated rows or missing values.

import pandas as pd

df = pd.read_csv("data/TSLA_clean.csv")
df["Date"] = pd.to_datetime(df["Date"])
df = df.set_index("Date").sort_index().drop_duplicates()

print("Duplicates:", df.duplicated().sum())
print("Missing:", df.isnull().sum().sum())

## Returns

Returns refer to the gain or loss made on an initial investment, often expressed as a percentage. We can use the generic **percentage change** formula here:

$$ (price_{end} - price_{start}) / price_{start} $$

We can apply this to close prices to calculate the simple daily return:

$$ (close price_{today} - close price_{yesterday}) / close price_{yesterday} $$

jan31_closing = df.loc["2024-01-31", "Close"]
jan30_closing = df.loc["2024-01-30", "Close"]

jan31_return = (jan31_closing - jan30_closing) / jan30_closing
print(f"Return on 31 Jan was {jan31_return:.2%}")
# f string: f"Return on 31 Jan was {jan31_return:.2%}", .2%s是formatting到2 decimal places

# f string小练习
name = "Lanxin"
print(f"Hi {name}!!")

This simple daily return expresses a loss in value of 2.24% from one day to the next. Notice we leave our return in decimal form, but when we output it we use `f-strings` and `:.2%` to display it as a percentage.

If we wanted to use the above approach to calculate daily returns for each day in our data set, it would take a long time. Let's see how we can use pandas `pct_change()` to make this sort of work easy, by applying our percentage change formula one column at a time.

df["Returns"] = df.Close.pct_change()
df
# .pct_change()是计算每一行与前一行的百分比变化

Notice how the first row in our data has a missing value **NaN** in the new daily return column. This is because our data doesn't have a close price for the day before it!

What to do with this missing value depends on what further analysis we want to do. If we want to carry out simple descriptive statistics like compute the mean, max, or standard deviation,  we can leave our missing value as NaN, because pandas will by default ignore NaNs when calculating these.

df.Returns.max() * 100
# 日收益率中的最大值的百分比，* 100是把小数改成百分比数
df.Returns.idxmax()
# 最大值所在的index/哪一天发生的
((1 + df.Returns).prod()-1)* 100
# 1 + df.Returns：把每一天的回报率变成“增长倍数”，如：2% → 1.02，-1% → 0.99
# .prod()：对所有这些增长倍数连乘,相当于计算整个期间的累计收益率；如：第一天涨 10%：→ 变成 1.10，第二天跌 5%：→ 变成 1.10 × 0.95 = 1.045
# -1:因为增长倍数从本金1开始，最后的是本金 × 增长倍数，所以减1得到实际收益率
# * 100：换成百分比
# 整段期间的累计收益率（百分比）


For more complex analyses though, we may want to drop or fill this value. Let's calculate cumulative returns for the period. Instead of comparing a given day with the day before it, cumulative returns compare a given day with the first day of the period, to indicate how our stock has performed since our initial investment.

We generally fill missing daily returns with a 0, which indicates no change with the day before.

Because we're doing cumulative multiplication, we'll add 1 to the closing price to get a growth factor, so we can compound the return over time. After calculating the cumulative product, we subtract 1 to get back to a return.

df ["Cumulative"] = (1 + df.Returns).cumprod()-1
# df ["Cumulative"]是生成了一个新的列
# .cumprod()是累计连乘，得到每一天的累计增长倍数的list
df

### Exercise: Bull or Bear?

Calculate the daily change in trading volume over the data frame. Then calculate the average change in trading volume over the period. Take the same approach we used for calculating daily returns, but considering `Volume` instead of `Close` price.

Then determine the trend in TSLA's stock for **Q1 2024**:

- Rising volume and increasing price might indicate a **bullish** trend (where the uptrend is backed by strong demand and could continue).

- Rising volume and decreasing price might indicate a **bearish** trend (where the downtrend is backed by strong selling pressure and could continue).

- Falling volume on price increase or decrease often indicate a trend is losing strength. It might suggest that momentum is waning and a price reversal is coming.

# 小测试
l = [1, 2, 3]
l.append (4)
l

l_copy = l.copy()
# 必须要这个code才能真的生成出copy，然后在copy上改，不会改动原数据l
l_copy.append(99)
l_copy


l
# verify一下

q1 = df.loc["2024-01-01" : "2024-03-31"].copy()
# subsetting q1 a copy, instead of using df directly

q1 ["VolChange"] = q1.Volume.pct_change()
print (f"Volume change is {q1.VolChange.mean()}")
print (f"Price change is {q1.Returns.mean()}")

# Rising volume and decreasing price, indicates a bearish trend


## Surges

Surges in price or trading volume can be helpful indicators for traders. One way to define a surge is as an increase on the day before by an amount higher than some defined threshold. That threshold is often defined as some number of standard deviations above the mean. Let's look at price surges of five standard deviations above the mean.

mean_return = df.Returns.mean()
threshold = mean_return + df.Returns.std() * 5
# 就如同题目说的，评判surge是看超过threshold的，threshold一般是比mean加上一定单位的std，这里是5个单位的std
condition = df.Returns > threshold
df[condition]

## Moving Averages

Moving averages are a different kind of indicator, one that smooths out small variations in trading data to give a better picture of the overall trend.

A Simple Moving Average (SMA) is one which averages out a price over a specific period. The average is "moving" because when a new day is considered in the period, the oldest date is discarded.

Moving averages can be *fast*, when they cover a short period, or *slow* when they consider a longer period. The longer the period, the more those small variations are smoothed out.

df["FastMA"] = df.Close.rolling(5).mean()
# rolling代表moving, (5)代表时间5天
df["SlowMA"] = df.Close.rolling(50).mean()
df
# 看表会发现前面的SlowMA都没数据，这是因为要50天了才能生成第一个数据

## Volatility

Volatility looks at the degree of variance in a stock, and can be helpful for determining risk. Periods of high standard deviation indicate higher volatility and may suggest a riskier investment.


df["Volatility"] = df.Returns.rolling(20).std()
df
# 用std展示volatility，rolling(20)是指每20天的std
