==> 11. Portfolios.ipynb <==
# Portfolio Management

A portfolio is a collection of financial investments. Let's see how we construct, analyse and optimise portfolios with Python. Like with every notebook, this one starts with a list of imports.

import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import seaborn as sns
import yfinance as yf

# 可以看到上面的package是以alphabet的顺序排的，对于python来讲，code能alphabetically排最好
# 更多的数据(dataframe)可以从WRDS摘取

plt.style.use("ggplot")

## Portfolio Formation

Today we'll form what is known as a **1/N portfolio**, with 5 stocks in it意思是有N个stock在portfolio每个都是equally weight. The 1/N refers to the total investment being divided equally among the N stocks that comprise the portfolio - which is 5 in our case today.

Depending on what kind of data we have, portfolio formation can involve different steps.

tickers = ["TSLA", "MSFT", "AMZN", "META", "NVDA"]

groups = []
for ticker in tickers:
# 从列表tickers中挨个“拿出元素”（只有“公司名”一个元素），皆放进变量ticker(可以是任何名字，ticker以方便)，然后去执行冒号后面的代码
    data = yf.download(ticker, start="2020-01-01", end="2020-12-31")
    # 对tickers list里面的每个ticker操作，让每个ticker的数据都等于yf上下载的数据
    data.columns = data.columns.droplevel(1)
    # 把每一列的名字弄成只有数据的名字，不带ticker名字
    data.insert(0, "Ticker", ticker)
    # 新建一列在最前面叫作"Ticker"，用ticker的名字做数据内容
    groups.append(data)

multi_df = pd.concat(groups)
# join data all together
multi_df

# run完之后会看到下面有个warning说auto_adjust是自动打开了；可以手动在data=的括号的最后加auto_adjust=False来让结果展示Adj Close和Close
# 这个adjust是来自于Corporate Actions，如Stock Split, Dividends, M&A，以真实反映投资回报。不adj的Close只是那天市场上收盘价，不包含后面发生的事
# 什么时候不adju的close有用？定trading strategy的时候

#### Multi-Stock Time Series Data (CRSP)
CRSP: Center for Research in Security Prices证券价格研究中心，提供股票历史金融数据，比Yahoo Finance的数据更好，用WRDS来access

Maybe instead of Yahoo! Finance, you took advantage of the [College WRDS access](https://www.imperial.ac.uk/admin-services/library/subject-support/business/crsp/), or obtained a CSV with your data in some other way.

Let's work with our panel data as before.

df = pd.read_csv("data/top_six_2020_2025.csv")

# 现在把panel变成cross sectional的
df_pivot = df.pivot(index="DlyCalDt", columns="Ticker", values="DlyClose")

df_pivot.index = pd.to_datetime(df_pivot.index, dayfirst=True)
df_pivot.sort_index(inplace=True)

# df_pivot.plot()，以看图展示这个pivot

# 现在要建立一个portfolio来放想要分析的5家公司，可以取前5个(01234)，作为portfolio:
# portfolio = df_pivot.columns[:5] 

# 还可以用loc：
# df_pivot.iloc[:, [0, 1, 3, 4]](要所有的行，不要2和5的公司列)，还可以：
df = df_pivot.loc[:, ["AMZN", "META", "MSFT", "NVDA", "AAPL"]]
df

# 现在需要define weights，我们现在要equal weights：
weights = np.ones(5) / 5
# np.ones(5):创建一个包含(5)个1的数组，对这个数组里的每个元素都除以5，每个得到equal weights
# 用这个公式的以快速适应元素数量的改变，如果是15个那就是np.ones(15) / 15
weights

## Calculating Portfolio Daily Returns

First, let's determine the returns of our portfolio. We'll need to start with the daily returns of each stock, and then use `np.dot()` to mutliply each column of returns against its weight in the portfolio.

returns = df.pct_change()
returns.dropna(inplace=True)
returns

# 那需要考虑每个stock的weights到每日return里面呢？
p_returns = np.dot(returns, weights)
# np.dot是dot product，做矩阵相乘，相当于sumproduct
print(f"portfolio's daily returns of the period is", p_returns)

p_volatility = p_returns.std()
print(f"portfolio's volatility is", p_volatility)

## Annualising Returns and Volatility

To provide a broader picture of our portfolio's performance, let's calculate the annualised portfolio return and volatility. These are important metrics for evaluating and discussing portfolios. They are also used frequently in more advanced financial analyses.

# annualise两组数据成为两个数据

TRASING_DAYS = 252
# how many trading days there are in an average year，根据不同market会有不同，一般american market是252天

annualised_returns = p_returns.mean() * TRASING_DAYS

annualised_volatility = p_volatility * np.sqrt(TRASING_DAYS)

print(
    f"""
    Annualised return is {annualised_returns}
    Annualised volatility is {annualised_volatility}
    """
    )
# 想输出的文字有多行，用三个引号可以让你直接写多行

## Sharpe Ratio

The Sharpe Ratio is a measure that helps investors understand the risk-adjusted return of an investment. A high Sharpe ratio indicates that the portfolio's returns are higher for each unit of risk taken on. In contrast, a lower Sharpe Ratio indicates a less favorable risk-reward trade-off, with the potential for lower returns relative to the amount of risk being assumed.

The Sharpe Ratio is calculated using annualised portfolio returns, portfolio volatility (as the measure of risk), and the *risk-free rate*. The risk-free rate is often derived from the yield of a theoretically risk-free investment, typically a government bond. The Sharpe Ratio is the average return earned in excess of the risk-free rate per unit of volatility or total risk.

### Exercise: Looking Sharpe

Calculate the Sharpe Ratio for our portfolio by applying the formula below.

$$ \text{Sharpe Ratio} = \frac{R_p - R_f}{\sigma_p} $$
这个公式告诉我们：how much return & risk, can be expected from the portfolio; 
高代表相对于risk来讲return较高，说明好

- $R_p$ is our annualised portfolio return
- $R_f$ is the risk-free rate
- $\sigma_p$ is our annualised volatility


RFR = 0.0438
# 课上那天老师现场去网上查的美国的risk free rate

sharpe = (annualised_returns - RFR) / annualised_volatility
print(f"Sharpe Ratio is: {sharpe}")
